requirements.txt
web3
openai
langchain-experimental
solders
solana
pyyaml
zstandard
requests
